employeeQuestions run properly with inquirer
