+++ lib/employee.js
@@ -1,22 +1,6 @@
-const Manager = require('./Manager')
-const Engineer = require('./Engineer')
-const Intern = require('./Intern')
-
-const employeeQuestions = [
-    {
-        type: 'input',
-        message: `Enter employee's name.`,
-        name: 'name',
-    }, {
-        type: 'input',
-        message: `Enter employee id.`,
-        name: 'name',
-    }, {
-        type: 'input',
-        message: `Enter employee's email address.`,
-        name: 'email',
-    } 
-]
+// const Manager = require('./Manager')
+// const Engineer = require('./Engineer')
+// const Intern = require('./Intern')
 
 class Employee {
     constructor(name, id, email) {
@@ -25,11 +9,17 @@ class Employee {
         this.email = email;
     }
 
-        getName() 
+        getName() {
+            return this.name;
+        }
+    
+        getId() {
+            return this.id;
+        }
 
-        getId()
-
-        getEmail()
+        getEmail() {
+            return this.email;
+        }
 
         getRole() {
             return 'Employee';
@@ -37,3 +27,4 @@ class Employee {
 }
 
 module.exports = Employee;
+
